+++ scarletio/core/traps/future_async_wrapper.py
@@ -10,7 +10,7 @@ from ..exceptions import InvalidStateError
 
 from .future import (
     FUTURE_STATE_CANCELLED, FUTURE_STATE_FINISHED, FUTURE_STATE_PENDING, FUTURE_STATE_RETRIEVED, Future,
-    get_future_state_name
+    get_exception_short_representation, get_future_state_name
 )
 
 
@@ -151,7 +151,7 @@ class FutureAsyncWrapper(Future):
                 repr_parts.append(reprlib.repr(self.result))
             else:
                 repr_parts.append(', exception=')
-                repr_parts.append(repr(exception))
+                repr_parts.append(get_exception_short_representation(exception))
         
         callbacks = self._callbacks
         limit = len(callbacks)